Fix a few more tests
